requirements.txt
albumentations==1.3.0
attr==0.3.2
cchardet==2.1.7
charset_normalizer==2.1.1
ConfigParser==5.3.0
contextlib2==21.6.0
cryptography==39.0.1
cupy==11.5.0
Cython==0.29.33
dl==0.1.0
docutils==0.19
filelock==3.9.0
grad_cam==1.4.6
html5lib==1.1
HTMLParser==0.0.2
ipython==8.10.0
ipywidgets==8.0.4
jax==0.4.3
Jinja2==3.1.2
jnius==1.1.0
keyring==23.13.1
lockfile==0.12.2
lxml==4.9.2
matplotlib==3.6.2
mock==5.0.1
mxnet==1.9.1
numpy==1.23.5
opencv_python_headless==4.7.0.68
pandas==1.5.2
Pillow==9.4.0
protobuf==4.21.12
pycocotools==2.0.6
pyOpenSSL==23.0.0
pytest==7.2.1
PyYAML==6.0
railroad==0.5.0
scikit_learn==1.2.1
Sphinx==6.1.3
torchvision==0.14.0+cu116
tornado==6.2
tqdm==4.64.1
trove_classifiers==2023.2.8
truststore==0.5.0
urllib3_secure_extra==0.1.0
win_inet_pton==1.1.0
xmlrpclib==1.0.1
